rename NftMint to NftMintPopup
